+++ examples/ackermann.c
@@ -5,9 +5,9 @@
 #define ack_IMPL(m, n)   M_uintMatchWithArgs(v(m), v(ack_), v(n))
 #define ack_S_IMPL(m, n) M_uintMatchWithArgs(v(n), v(ack_S_), v(m))
 
-#define ack_Z_IMPL(n)      M_uintInc(v(n))
+#define ack_Z_IMPL(n)      M_inc(v(n))
 #define ack_S_Z_IMPL(m)    ack(v(m), v(1))
-#define ack_S_S_IMPL(n, m) ack(v(m), ack(M_uintInc(v(m)), v(n)))
+#define ack_S_S_IMPL(n, m) ack(v(m), ack(M_inc(v(m)), v(n)))
 
 M_assertEq(ack(v(0), v(0)), v(1));
 M_assertEq(ack(v(0), v(1)), v(2));