+++ examples/factorial.c
@@ -4,7 +4,7 @@
 
 #define factorial_IMPL(n)   M_uintMatch(v(n), v(factorial_))
 #define factorial_Z_IMPL()  v(1)
-#define factorial_S_IMPL(n) M_uintMul(M_uintInc(v(n)), factorial(v(n)))
+#define factorial_S_IMPL(n) M_mul(M_inc(v(n)), factorial(v(n)))
 
 M_assertEq(factorial(v(0)), v(1));
 M_assertEq(factorial(v(1)), v(1));