+++ include/metalang99/uint.h
@@ -28,12 +28,12 @@
  * #include <metalang99/uint.h>
  *
  * // 6
- * M_uintInc(v(5))
+ * M_inc(v(5))
  * @endcode
  *
  * @note If @p x is 255, the result is 0.
  */
-#define METALANG99_uintInc(x) METALANG99_call(METALANG99_uintInc, x)
+#define METALANG99_inc(x) METALANG99_call(METALANG99_inc, x)
 
 /**
  * \f$x - 1\f$
@@ -44,12 +44,12 @@
  * #include <metalang99/uint.h>
  *
  * // 4
- * M_uintDec(v(5))
+ * M_dec(v(5))
  * @endcode
  *
  * @note If @p x is 0, the result is 255.
  */
-#define METALANG99_uintDec(x) METALANG99_call(METALANG99_uintDec, x)
+#define METALANG99_dec(x) METALANG99_call(METALANG99_dec, x)
 
 /**
  * Matches @p x to the two cases: if it is zero or positive.
@@ -141,13 +141,13 @@
  * #include <metalang99/uint.h>
  *
  * // 1
- * M_uintGreater(v(8), v(3))
+ * M_greater(v(8), v(3))
  *
  * // 0
- * M_uintGreater(v(3), v(8))
+ * M_greater(v(3), v(8))
  * @endcode
  */
-#define METALANG99_uintGreater(x, y) METALANG99_call(METALANG99_uintGreater, x, y)
+#define METALANG99_greater(x, y) METALANG99_call(METALANG99_greater, x, y)
 
 /**
  * \f$x \geq y\f$
@@ -158,13 +158,13 @@
  * #include <metalang99/uint.h>
  *
  * // 1
- * M_uintGreaterEq(v(8), v(8))
+ * M_greaterEq(v(8), v(8))
  *
  * // 0
- * M_uintGreaterEq(v(3), v(8))
+ * M_greaterEq(v(3), v(8))
  * @endcode
  */
-#define METALANG99_uintGreaterEq(x, y) METALANG99_call(METALANG99_uintGreaterEq, x, y)
+#define METALANG99_greaterEq(x, y) METALANG99_call(METALANG99_greaterEq, x, y)
 
 /**
  * \f$x < y\f$
@@ -175,13 +175,13 @@
  * #include <metalang99/uint.h>
  *
  * // 1
- * M_uintLesser(v(3), v(8))
+ * M_lesser(v(3), v(8))
  *
  * // 0
- * M_uintLesser(v(8), v(3))
+ * M_lesser(v(8), v(3))
  * @endcode
  */
-#define METALANG99_uintLesser(x, y) METALANG99_call(METALANG99_uintLesser, x, y)
+#define METALANG99_lesser(x, y) METALANG99_call(METALANG99_lesser, x, y)
 
 /**
  * \f$x \leq y\f$
@@ -192,13 +192,13 @@
  * #include <metalang99/uint.h>
  *
  * // 1
- * M_uintLesserEq(v(8), v(8))
+ * M_lesserEq(v(8), v(8))
  *
  * // 0
- * M_uintLesserEq(v(8), v(3))
+ * M_lesserEq(v(8), v(3))
  * @endcode
  */
-#define METALANG99_uintLesserEq(x, y) METALANG99_call(METALANG99_uintLesserEq, x, y)
+#define METALANG99_lesserEq(x, y) METALANG99_call(METALANG99_lesserEq, x, y)
 
 /**
  * \f$x + y\f$
@@ -209,10 +209,10 @@
  * #include <metalang99/uint.h>
  *
  * // 11
- * M_uintAdd(v(5), v(6))
+ * M_add(v(5), v(6))
  * @endcode
  */
-#define METALANG99_uintAdd(x, y) METALANG99_call(METALANG99_uintAdd, x, y)
+#define METALANG99_add(x, y) METALANG99_call(METALANG99_add, x, y)
 
 /**
  * \f$x - y\f$
@@ -223,10 +223,10 @@
  * #include <metalang99/uint.h>
  *
  * // 6
- * M_uintSub(v(11), v(5))
+ * M_sub(v(11), v(5))
  * @endcode
  */
-#define METALANG99_uintSub(x, y) METALANG99_call(METALANG99_uintSub, x, y)
+#define METALANG99_sub(x, y) METALANG99_call(METALANG99_sub, x, y)
 
 /**
  * \f$x * y\f$
@@ -237,10 +237,10 @@
  * #include <metalang99/uint.h>
  *
  * // 12
- * M_uintMul(v(3), v(4))
+ * M_mul(v(3), v(4))
  * @endcode
  */
-#define METALANG99_uintMul(x, y) METALANG99_call(METALANG99_uintMul, x, y)
+#define METALANG99_mul(x, y) METALANG99_call(METALANG99_mul, x, y)
 
 /**
  * \f$\frac{x}{y}\f$
@@ -251,15 +251,15 @@
  * #include <metalang99/uint.h>
  *
  * // 3
- * M_uintDiv(v(12), v(4))
+ * M_div(v(12), v(4))
  * @endcode
  *
  * @note A compile-time error if \f$\frac{x}{y}\f$ is not an unsigned integer.
  */
-#define METALANG99_uintDiv(x, y) METALANG99_call(METALANG99_uintDiv, x, y)
+#define METALANG99_div(x, y) METALANG99_call(METALANG99_div, x, y)
 
 /**
- * Like #METALANG99_uintDiv but returns #METALANG99_nothing is @p x is not divisible by @p y,
+ * Like #METALANG99_div but returns #METALANG99_nothing is @p x is not divisible by @p y,
  * otherwise, #METALANG99_just.
  *
  * # Examples
@@ -268,16 +268,16 @@
  * #include <metalang99/uint.h>
  *
  * // M_just(3)
- * M_uintDivChecked(v(12), v(4))
+ * M_divChecked(v(12), v(4))
  *
  * // M_nothing()
- * M_uintDivChecked(v(14), v(5))
+ * M_divChecked(v(14), v(5))
  *
  * // M_nothing()
- * M_uintDivChecked(v(1), v(0))
+ * M_divChecked(v(1), v(0))
  * @endcode
  */
-#define METALANG99_uintDivChecked(x, y) METALANG99_call(METALANG99_uintDivChecked, x, y)
+#define METALANG99_divChecked(x, y) METALANG99_call(METALANG99_divChecked, x, y)
 
 /**
  * Computes the remainder of division.
@@ -288,12 +288,12 @@
  * #include <metalang99/uint.h>
  *
  * // 2
- * M_uintMod(v(8), v(3))
+ * M_mod(v(8), v(3))
  * @endcode
  *
  * @note A compile-time error if @p y is 0.
  */
-#define METALANG99_uintMod(x, y) METALANG99_call(METALANG99_uintMod, x, y)
+#define METALANG99_mod(x, y) METALANG99_call(METALANG99_mod, x, y)
 
 /**
  * \f$x + y + z\f$
@@ -304,10 +304,10 @@
  * #include <metalang99/uint.h>
  *
  * // 15
- * M_uintAdd3(v(1), v(6), v(8))
+ * M_add3(v(1), v(6), v(8))
  * @endcode
  */
-#define METALANG99_uintAdd3(x, y, z) METALANG99_call(METALANG99_uintAdd3, x, y, z)
+#define METALANG99_add3(x, y, z) METALANG99_call(METALANG99_add3, x, y, z)
 
 /**
  * \f$x - y - z\f$
@@ -318,10 +318,10 @@
  * #include <metalang99/uint.h>
  *
  * // 3
- * M_uintSub3(v(8), v(2), v(3))
+ * M_sub3(v(8), v(2), v(3))
  * @endcode
  */
-#define METALANG99_uintSub3(x, y, z) METALANG99_call(METALANG99_uintSub3, x, y, z)
+#define METALANG99_sub3(x, y, z) METALANG99_call(METALANG99_sub3, x, y, z)
 
 /**
  * \f$x * y * z\f$
@@ -332,10 +332,10 @@
  * #include <metalang99/uint.h>
  *
  * // 24
- * M_uintMul3(v(2), v(3), v(4))
+ * M_mul3(v(2), v(3), v(4))
  * @endcode
  */
-#define METALANG99_uintMul3(x, y, z) METALANG99_call(METALANG99_uintMul3, x, y, z)
+#define METALANG99_mul3(x, y, z) METALANG99_call(METALANG99_mul3, x, y, z)
 
 /**
  * \f$\frac{(\frac{x}{y})}{z}\f$
@@ -346,12 +346,12 @@
  * #include <metalang99/uint.h>
  *
  * // 5
- * M_uintDiv(v(30), v(3), v(2))
+ * M_div(v(30), v(3), v(2))
  * @endcode
  *
  * @note A compile-time error if \f$\frac{(\frac{x}{y})}{z}\f$ is not an unsigned integer.
  */
-#define METALANG99_uintDiv3(x, y, z) METALANG99_call(METALANG99_uintDiv3, x, y, z)
+#define METALANG99_div3(x, y, z) METALANG99_call(METALANG99_div3, x, y, z)
 
 /**
  * \f$min(x, y)\f$
@@ -362,10 +362,10 @@
  * #include <metalang99/uint.h>
  *
  * // 5
- * M_uintMin(v(5), v(7))
+ * M_min(v(5), v(7))
  * @endcode
  */
-#define METALANG99_uintMin(x, y) METALANG99_call(METALANG99_uintMin, x, y)
+#define METALANG99_min(x, y) METALANG99_call(METALANG99_min, x, y)
 
 /**
  * \f$max(x, y)\f$
@@ -376,10 +376,10 @@
  * #include <metalang99/uint.h>
  *
  * // 7
- * M_uintMax(v(5), v(7))
+ * M_max(v(5), v(7))
  * @endcode
  */
-#define METALANG99_uintMax(x, y) METALANG99_call(METALANG99_uintMax, x, y)
+#define METALANG99_max(x, y) METALANG99_call(METALANG99_max, x, y)
 
 /**
  * The plain version of #METALANG99_uintEq.
@@ -392,14 +392,14 @@
 #define METALANG99_uintNeqPlain(x, y) METALANG99_notPlain(METALANG99_uintEqPlain(x, y))
 
 /**
- * The plain version of #METALANG99_uintIncPlain.
+ * The plain version of #METALANG99_incPlain.
  */
-#define METALANG99_uintIncPlain(x) METALANG99_PRIV_uintInc(x)
+#define METALANG99_incPlain(x) METALANG99_PRIV_inc(x)
 
 /**
- * The plain version of #METALANG99_uintDecPlain.
+ * The plain version of #METALANG99_decPlain.
  */
-#define METALANG99_uintDecPlain(x) METALANG99_PRIV_uintDec(x)
+#define METALANG99_decPlain(x) METALANG99_PRIV_dec(x)
 // }
 
 #ifndef DOXYGEN_IGNORE
@@ -410,198 +410,178 @@
     METALANG99_PRIV_IF(                                                                            \
         METALANG99_PRIV_uintEq(x, 0),                                                              \
         METALANG99_callTrivial(matcher##Z, ),                                                      \
-        METALANG99_callTrivial(matcher##S, METALANG99_PRIV_uintDec(x)))
+        METALANG99_callTrivial(matcher##S, METALANG99_PRIV_dec(x)))
 
 #define METALANG99_uintMatchWithArgs_IMPL(x, matcher, ...)                                         \
     METALANG99_PRIV_IF(                                                                            \
         METALANG99_PRIV_uintEq(x, 0),                                                              \
         METALANG99_callTrivial(matcher##Z, __VA_ARGS__),                                           \
-        METALANG99_callTrivial(matcher##S, METALANG99_PRIV_uintDec(x), __VA_ARGS__))
+        METALANG99_callTrivial(matcher##S, METALANG99_PRIV_dec(x), __VA_ARGS__))
 
-#define METALANG99_uintEq_IMPL(x, y)  v(METALANG99_uintEqPlain(x, y))
-#define METALANG99_uintNeq_IMPL(x, y) v(METALANG99_uintNeqPlain(x, y))
-#define METALANG99_uintInc_IMPL(x)    v(METALANG99_uintIncPlain(x))
-#define METALANG99_uintDec_IMPL(x)    v(METALANG99_uintDecPlain(x))
+#define METALANG99_uintEq_IMPL(x, y)   v(METALANG99_uintEqPlain(x, y))
+#define METALANG99_uintNeq_IMPL(x, y)  v(METALANG99_uintNeqPlain(x, y))
+#define METALANG99_inc_IMPL(x)         v(METALANG99_incPlain(x))
+#define METALANG99_dec_IMPL(x)         v(METALANG99_decPlain(x))
+#define METALANG99_greater_IMPL(x, y)  METALANG99_lesser_IMPL(y, x)
+#define METALANG99_lesserEq_IMPL(x, y) METALANG99_greaterEq_IMPL(y, x)
 
-#define METALANG99_uintGreater_IMPL(x, y)  METALANG99_uintLesser_IMPL(y, x)
-#define METALANG99_uintLesserEq_IMPL(x, y) METALANG99_uintGreaterEq_IMPL(y, x)
+#define METALANG99_greaterEq_IMPL(x, y)                                                            \
+    METALANG99_PRIV_IF(                                                                            \
+        METALANG99_PRIV_uintEq(x, y),                                                              \
+        v(METALANG99_true),                                                                        \
+        METALANG99_greater_IMPL(x, y))
 
-#define METALANG99_uintGreaterEq_IMPL(x, y)                                                        \
-    METALANG99_callTrivial(                                                                        \
-        METALANG99_PRIV_IF(                                                                        \
-            METALANG99_PRIV_uintEq(x, y),                                                          \
-            METALANG99_PRIV_CONST_TRUE,                                                            \
-            METALANG99_uintGreater),                                                               \
-        x,                                                                                         \
-        y)
-
-// METALANG99_uintLesser_IMPL {
-#define METALANG99_uintLesser_IMPL(x, y)                                                           \
-    METALANG99_callTrivial(                                                                        \
-        METALANG99_PRIV_IF(                                                                        \
-            METALANG99_PRIV_uintEq(y, 0),                                                          \
-            METALANG99_PRIV_CONST_FALSE,                                                           \
-            METALANG99_PRIV_uintLesserProgress),                                                   \
-        x,                                                                                         \
-        y)
-#define METALANG99_PRIV_uintLesserProgress_IMPL(x, y)                                              \
-    METALANG99_callTrivial(                                                                        \
-        METALANG99_PRIV_IF(                                                                        \
-            METALANG99_PRIV_uintEq(x, METALANG99_PRIV_uintDec(y)),                                 \
-            METALANG99_PRIV_CONST_TRUE,                                                            \
-            METALANG99_uintLesser),                                                                \
-        x,                                                                                         \
-        METALANG99_PRIV_uintDec(y))
-// }
-
-// METALANG99_uintAdd_IMPL {
-#define METALANG99_uintAdd_IMPL(x, y)                                                              \
-    METALANG99_callTrivial(                                                                        \
-        METALANG99_PRIV_IF(                                                                        \
-            METALANG99_PRIV_uintEq(y, 0),                                                          \
-            METALANG99_const,                                                                      \
-            METALANG99_PRIV_uintAddProgress),                                                      \
-        x,                                                                                         \
-        y)
-#define METALANG99_PRIV_uintAddProgress_IMPL(x, y)                                                 \
-    METALANG99_callTrivial(                                                                        \
-        METALANG99_uintAdd,                                                                        \
-        METALANG99_PRIV_uintInc(x),                                                                \
-        METALANG99_PRIV_uintDec(y))
-// }
-
-// METALANG99_uintSub_IMPL {
-#define METALANG99_uintSub_IMPL(x, y)                                                              \
-    METALANG99_callTrivial(                                                                        \
-        METALANG99_PRIV_IF(                                                                        \
-            METALANG99_PRIV_uintEq(y, 0),                                                          \
-            METALANG99_const,                                                                      \
-            METALANG99_PRIV_uintSubProgress),                                                      \
-        x,                                                                                         \
-        y)
-#define METALANG99_PRIV_uintSubProgress_IMPL(x, y)                                                 \
-    METALANG99_uintSub_IMPL(METALANG99_PRIV_uintDec(x), METALANG99_PRIV_uintDec(y))
-// }
-
-// METALANG99_uintMul_IMPL {
-#define METALANG99_uintMul_IMPL(x, y)                                                              \
-    METALANG99_callTrivial(                                                                        \
-        METALANG99_PRIV_IF(                                                                        \
-            METALANG99_PRIV_uintEq(y, 0),                                                          \
-            METALANG99_PRIV_UINT_CONST_0,                                                          \
-            METALANG99_PRIV_uintMulProgress),                                                      \
-        x,                                                                                         \
-        y)
-#define METALANG99_PRIV_uintMulProgress_IMPL(x, y)                                                 \
-    METALANG99_uintAdd(v(x), METALANG99_uintMul_IMPL(x, METALANG99_PRIV_uintDec(y)))
-// }
-
-// METALANG99_uintMod_IMPL {
-#define METALANG99_uintMod_IMPL(x, y)                                                              \
+// METALANG99_lesser_IMPL {
+#define METALANG99_lesser_IMPL(x, y)                                                               \
     METALANG99_PRIV_IF(                                                                            \
         METALANG99_PRIV_uintEq(y, 0),                                                              \
-        METALANG99_fatal(METALANG99_uintMod, modulo by 0),                                         \
-        METALANG99_callTrivial(METALANG99_PRIV_uintModAux, x, y, 0))
+        v(METALANG99_false),                                                                       \
+        METALANG99_callTrivial(METALANG99_PRIV_lesserProgress, x, y))
+#define METALANG99_PRIV_lesserProgress_IMPL(x, y)                                                  \
+    METALANG99_PRIV_IF(                                                                            \
+        METALANG99_PRIV_uintEq(x, METALANG99_PRIV_dec(y)),                                         \
+        v(METALANG99_true),                                                                        \
+        METALANG99_callTrivial(METALANG99_lesser, x, METALANG99_PRIV_dec(y)))                      \
+// }
 
-#define METALANG99_PRIV_uintModAux_IMPL(x, y, acc)                                                 \
+// METALANG99_add_IMPL {
+#define METALANG99_add_IMPL(x, y)                                                                  \
+    METALANG99_callTrivial(                                                                        \
+        METALANG99_PRIV_IF(                                                                        \
+            METALANG99_PRIV_uintEq(y, 0),                                                          \
+            METALANG99_const,                                                                      \
+            METALANG99_PRIV_addProgress),                                                          \
+        x,                                                                                         \
+        y)
+#define METALANG99_PRIV_addProgress_IMPL(x, y)                                                     \
+    METALANG99_add_IMPL(METALANG99_PRIV_inc(x), METALANG99_PRIV_dec(y))
+// }
+
+// METALANG99_sub_IMPL {
+#define METALANG99_sub_IMPL(x, y)                                                                  \
+    METALANG99_callTrivial(                                                                        \
+        METALANG99_PRIV_IF(                                                                        \
+            METALANG99_PRIV_uintEq(y, 0),                                                          \
+            METALANG99_const,                                                                      \
+            METALANG99_PRIV_subProgress),                                                          \
+        x,                                                                                         \
+        y)
+#define METALANG99_PRIV_subProgress_IMPL(x, y)                                                     \
+    METALANG99_sub_IMPL(METALANG99_PRIV_dec(x), METALANG99_PRIV_dec(y))
+// }
+
+// METALANG99_mul_IMPL {
+#define METALANG99_mul_IMPL(x, y)                                                                  \
+    METALANG99_PRIV_IF(                                                                            \
+        METALANG99_PRIV_uintEq(y, 0),                                                              \
+        v(0),                                                                                      \
+        METALANG99_callTrivial(METALANG99_PRIV_mulProgress, x, y))
+#define METALANG99_PRIV_mulProgress_IMPL(x, y)                                                     \
+    METALANG99_add(v(x), METALANG99_mul_IMPL(x, METALANG99_PRIV_dec(y)))
+// }
+
+// METALANG99_mod_IMPL {
+#define METALANG99_mod_IMPL(x, y)                                                                  \
+    METALANG99_PRIV_IF(                                                                            \
+        METALANG99_PRIV_uintEq(y, 0),                                                              \
+        METALANG99_fatal(METALANG99_mod, modulo by 0),                                             \
+        METALANG99_callTrivial(METALANG99_PRIV_modAux, x, y, 0))
+
+#define METALANG99_PRIV_modAux_IMPL(x, y, acc)                                                     \
     METALANG99_appl(                                                                               \
         METALANG99_if(                                                                             \
-            METALANG99_isJust(METALANG99_callTrivial(METALANG99_uintDivChecked, x, y)),            \
-            METALANG99_callTrivial(METALANG99_appl, METALANG99_const, acc),                        \
-            METALANG99_callTrivial(METALANG99_appl2, METALANG99_PRIV_uintModProgress, x, y)),      \
+            METALANG99_isJust(METALANG99_divChecked_IMPL(x, y)),                                   \
+            METALANG99_appl_IMPL(METALANG99_const, acc),                                           \
+            METALANG99_appl2_IMPL(METALANG99_PRIV_modProgress, x, y)),                             \
         v(acc))
 
-#define METALANG99_PRIV_uintModProgress_IMPL(x, y, acc)                                            \
+#define METALANG99_PRIV_modProgress_IMPL(x, y, acc)                                                \
     METALANG99_appl(                                                                               \
         METALANG99_PRIV_IF(                                                                        \
             METALANG99_PRIV_uintEq(x, 0),                                                          \
-            METALANG99_callTrivial(METALANG99_appl, METALANG99_const, acc),                        \
-            METALANG99_callTrivial(METALANG99_appl2, METALANG99_PRIV_uintModProgressAux, x, y)),   \
+            METALANG99_appl_IMPL(METALANG99_const, acc),                                           \
+            METALANG99_appl2_IMPL(METALANG99_PRIV_modProgressAux, x, y)),                          \
         v(acc))
 
-#define METALANG99_PRIV_uintModProgressAux_IMPL(x, y, acc)                                         \
-    METALANG99_callTrivial(                                                                        \
-        METALANG99_PRIV_uintModAux,                                                                \
-        METALANG99_PRIV_uintDec(x),                                                                \
-        y,                                                                                         \
-        METALANG99_PRIV_uintInc(acc))
+#define METALANG99_PRIV_modProgressAux_IMPL(x, y, acc)                                             \
+    METALANG99_PRIV_modAux_IMPL(METALANG99_PRIV_dec(x), y, METALANG99_PRIV_inc(acc))
 // }
 
-#define METALANG99_uintAdd3_IMPL(x, y, z) METALANG99_uintAdd(METALANG99_uintAdd_IMPL(x, y), v(z))
-#define METALANG99_uintSub3_IMPL(x, y, z) METALANG99_uintSub(METALANG99_uintSub_IMPL(x, y), v(z))
-#define METALANG99_uintMul3_IMPL(x, y, z) METALANG99_uintMul(METALANG99_uintMul_IMPL(x, y), v(z))
-#define METALANG99_uintDiv3_IMPL(x, y, z) METALANG99_uintDiv(METALANG99_uintDiv_IMPL(x, y), v(z))
+#define METALANG99_add3_IMPL(x, y, z) METALANG99_add(METALANG99_add_IMPL(x, y), v(z))
+#define METALANG99_sub3_IMPL(x, y, z) METALANG99_sub(METALANG99_sub_IMPL(x, y), v(z))
+#define METALANG99_mul3_IMPL(x, y, z) METALANG99_mul(METALANG99_mul_IMPL(x, y), v(z))
+#define METALANG99_div3_IMPL(x, y, z) METALANG99_div(METALANG99_div_IMPL(x, y), v(z))
 
-#define METALANG99_uintMin_IMPL(x, y)                                                              \
-    METALANG99_call(METALANG99_if, METALANG99_uintLesser_IMPL(x, y), v(x, y))
-#define METALANG99_uintMax_IMPL(x, y)                                                              \
-    METALANG99_call(METALANG99_if, METALANG99_uintLesser_IMPL(x, y), v(y, x))
+#define METALANG99_min_IMPL(x, y)                                                                  \
+    METALANG99_call(METALANG99_if, METALANG99_lesser_IMPL(x, y), v(x, y))
+#define METALANG99_max_IMPL(x, y)                                                                  \
+    METALANG99_call(METALANG99_if, METALANG99_lesser_IMPL(x, y), v(y, x))
 
-#define METALANG99_PRIV_UINT_CONST_0_IMPL(...) v(0)
+#define METALANG99_PRIV_uintConst0_IMPL(...) v(0)
 // }
 
 // Arity specifiers {
-#define METALANG99_uintInc_ARITY           1
-#define METALANG99_uintDec_ARITY           1
+#define METALANG99_inc_ARITY               1
+#define METALANG99_dec_ARITY               1
 #define METALANG99_uintMatch_ARITY         2
 #define METALANG99_uintMatchWithArgs_ARITY 3
 #define METALANG99_uintEq_ARITY            2
 #define METALANG99_uintNeq_ARITY           2
-#define METALANG99_uintGreater_ARITY       2
-#define METALANG99_uintGreaterEq_ARITY     2
-#define METALANG99_uintLesser_ARITY        2
-#define METALANG99_uintLesserEq_ARITY      2
-#define METALANG99_uintAdd_ARITY           2
-#define METALANG99_uintSub_ARITY           2
-#define METALANG99_uintMul_ARITY           2
-#define METALANG99_uintDiv_ARITY           2
-#define METALANG99_uintDivChecked_ARITY    2
-#define METALANG99_uintMod_ARITY           2
-#define METALANG99_uintAdd3_ARITY          3
-#define METALANG99_uintSub3_ARITY          3
-#define METALANG99_uintMul3_ARITY          3
-#define METALANG99_uintDiv3_ARITY          3
-#define METALANG99_uintMin_ARITY           2
-#define METALANG99_uintMax_ARITY           2
+#define METALANG99_greater_ARITY           2
+#define METALANG99_greaterEq_ARITY         2
+#define METALANG99_lesser_ARITY            2
+#define METALANG99_lesserEq_ARITY          2
+#define METALANG99_add_ARITY               2
+#define METALANG99_sub_ARITY               2
+#define METALANG99_mul_ARITY               2
+#define METALANG99_div_ARITY               2
+#define METALANG99_divChecked_ARITY        2
+#define METALANG99_mod_ARITY               2
+#define METALANG99_add3_ARITY              3
+#define METALANG99_sub3_ARITY              3
+#define METALANG99_mul3_ARITY              3
+#define METALANG99_div3_ARITY              3
+#define METALANG99_min_ARITY               2
+#define METALANG99_max_ARITY               2
 
-#define METALANG99_PRIV_uintLesserProgress_ARITY 2
-#define METALANG99_PRIV_uintAddProgress_ARITY    2
-#define METALANG99_PRIV_uintSubProgress_ARITY    2
-#define METALANG99_PRIV_uintMulProgress_ARITY    2
-#define METALANG99_PRIV_uintModProgress_ARITY    3
-#define METALANG99_PRIV_uintModProgressAux_ARITY 3
+#define METALANG99_PRIV_lesserProgress_ARITY 2
+#define METALANG99_PRIV_addProgress_ARITY    2
+#define METALANG99_PRIV_subProgress_ARITY    2
+#define METALANG99_PRIV_mulProgress_ARITY    2
+#define METALANG99_PRIV_modProgress_ARITY    3
+#define METALANG99_PRIV_modProgressAux_ARITY 3
 // }
 
 // Aliases {
 #ifndef METALANG99_NO_SMALL_PREFIX
 
-#define M_uintInc           METALANG99_uintInc
-#define M_uintDec           METALANG99_uintDec
+#define M_inc               METALANG99_inc
+#define M_dec               METALANG99_dec
 #define M_uintMatch         METALANG99_uintMatch
 #define M_uintMatchWithArgs METALANG99_uintMatchWithArgs
 #define M_uintEq            METALANG99_uintEq
 #define M_uintNeq           METALANG99_uintNeq
-#define M_uintGreater       METALANG99_uintGreater
-#define M_uintGreaterEq     METALANG99_uintGreaterEq
-#define M_uintLesser        METALANG99_uintLesser
-#define M_uintLesserEq      METALANG99_uintLesserEq
-#define M_uintAdd           METALANG99_uintAdd
-#define M_uintSub           METALANG99_uintSub
-#define M_uintMul           METALANG99_uintMul
-#define M_uintDiv           METALANG99_uintDiv
-#define M_uintDivChecked    METALANG99_uintDivChecked
-#define M_uintMod           METALANG99_uintMod
-#define M_uintAdd3          METALANG99_uintAdd3
-#define M_uintSub3          METALANG99_uintSub3
-#define M_uintMul3          METALANG99_uintMul3
-#define M_uintDiv3          METALANG99_uintDiv3
-#define M_uintMin           METALANG99_uintMin
-#define M_uintMax           METALANG99_uintMax
+#define M_greater           METALANG99_greater
+#define M_greaterEq         METALANG99_greaterEq
+#define M_lesser            METALANG99_lesser
+#define M_lesserEq          METALANG99_lesserEq
+#define M_add               METALANG99_add
+#define M_sub               METALANG99_sub
+#define M_mul               METALANG99_mul
+#define M_div               METALANG99_div
+#define M_divChecked        METALANG99_divChecked
+#define M_mod               METALANG99_mod
+#define M_add3              METALANG99_add3
+#define M_sub3              METALANG99_sub3
+#define M_mul3              METALANG99_mul3
+#define M_div3              METALANG99_div3
+#define M_min               METALANG99_min
+#define M_max               METALANG99_max
 
 #define M_uintEqPlain  METALANG99_uintEqPlain
 #define M_uintNeqPlain METALANG99_uintNeqPlain
-#define M_uintIncPlain METALANG99_uintIncPlain
-#define M_uintDecPlain METALANG99_uintDecPlain
+#define M_incPlain     METALANG99_incPlain
+#define M_decPlain     METALANG99_decPlain
 
 #endif // METALANG99_NO_SMALL_PREFIX
 // }