+++ CHANGELOG.md
@@ -29,6 +29,19 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
    - `M_isParenthesised` -> `M_isParenthesized`
    - `M_isUnparenthesised` -> `M_isUnparenthesized`
    - `M_parenthesisedVariadics(Head|Tail)` -> `M_parenthesizedVariadics(Head|Tail)`
+ - Shorten functions on unsigned integers:
+   - `M_uintInc(Plain)` -> `M_inc(Plain)`
+   - `M_uintDec(Plain)` -> `M_dec(Plain)`
+   - `M_uintAdd(2|3)` -> `M_add(2|3)`
+   - `M_uintSub(2|3)` -> `M_sub(2|3)`
+   - `M_uintMul(2|3)` -> `M_mul(2|3)`
+   - `M_uintDiv(2|3)` -> `M_div(2|3)`
+   - `M_uintDivChecked` -> `M_divChecked`
+   - `M_uintLesser(Eq)` -> `M_lesser(Eq)`
+   - `M_uintGreater(Eq)` -> `M_greater(Eq)`
+   - `M_uintMod` -> `M_mod`
+   - `M_uintMin` -> `M_min`
+   - `M_uintMax` -> `M_max`
 
 ### Fixed